+++ hume-editor/src/editor/mod.rs
@@ -866,4 +866,4 @@ impl Editor {
 }
 
 #[cfg(test)]
-mod tests;
+pub(crate) mod tests;